requirements.txt
Django==4.0.3
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0